Move navbar component to own folder, dynamically change width depending on search or not
